add logic for adding ScheduleGaps
ScheduleGaps are gaps between two ScheduleBlocks or the start/end
of the schedule day and the first/last ScheduleBlock.
Like ScheduleBlocks they have an end time, start time and a duration.
They help users to add new a ScheduleBlock between two ScheduleBlocks
or at the start/end of the schedule