fix bug: undoing old deletion of category
Description of bug:
There are two categories A (parent) and B (child). When category B is deleted, it is marked for deletion. If category A is subsequently deleted and the deletion is then reversed, the deletion of category B is also reversed although only the last deletion (of category A) should be reversed.