fix: use old version of navigation library
because of weird bug: when adding todos or categories for a schedule block with the picker and then immediately saving the schedule block the todos/categories are not saved.
This happens because the add button of the picker screen gets called twice, resulting in calling navController.popBackStack twice too. The second call of popBackStack occurs before the save event can be processed, resulting in the scheduleBlock not being saved.
This may happen because of animations introduced with the version of the navigation library.
Using the previous version of the library the issue no longer occurs.